Namespace reading works
