+++ src/hope_static_pass.h
@@ -122,7 +122,8 @@ private:
         void resolve();
 
     private:
-        ParseNode resolveStmt(size_t pn) noexcept;
+        ParseNode resolveStmt(ParseNode pn) noexcept;
+        ParseNode resolveLValue(ParseNode pn) noexcept;
         Type fillDefaultsAndInstantiate(ParseNode call_node, CallSignature sig);
         ParseNode resolveExprTop(size_t pn, size_t rows_expected = 0, size_t cols_expected = 0);
         ParseNode resolveExpr(size_t pn, size_t rows_expected = 0, size_t cols_expected = 0) noexcept;
@@ -149,6 +150,7 @@ private:
         ParseNode resolveZeroMatrix(ParseNode pn);
         ParseNode resolveLimit(ParseNode pn);
         ParseNode resolveDefiniteIntegral(ParseNode pn);
+        ParseNode resolveScopeAccess(ParseNode pn);
         ParseNode copyChildProperties(ParseNode pn) noexcept;
         ParseNode enforceScalar(ParseNode pn);
         ParseNode enforceZero(ParseNode pn);
@@ -174,6 +176,8 @@ private:
         std::vector<CachedInfo> old_val_cap;
         std::vector<CachedInfo> old_ref_cap;
         std::vector<CachedInfo> old_args;
+
+        size_t closureDepth() const noexcept { return return_types.size(); }
 };
 
 }
