+++ src/hope_symbol_table.h
@@ -128,6 +128,22 @@ public:
     #ifndef NDEBUG
     void verifyIdentifier(ParseNode pn) const noexcept;
     #endif
+
+    struct StoredScopeKey {
+        ParseNode symbol_root;
+        Typeset::Selection sel;
+        StoredScopeKey(ParseNode symbol_root, const Typeset::Selection& sel) noexcept : symbol_root(symbol_root), sel(sel) {}
+        bool operator==(const StoredScopeKey& other) const noexcept {
+            return symbol_root == other.symbol_root && sel == other.sel; }
+    };
+    struct HashStoredScopeKey {
+        std::size_t operator()(const StoredScopeKey& s) const {
+            return s.symbol_root ^ std::hash<Typeset::Selection>()(s.sel);
+        }
+    };
+    HOPE_UNORDERED_MAP<StoredScopeKey, size_t, HashStoredScopeKey> stored_scopes;
+
+    void resolveReference(ParseNode pn, size_t sym_id, size_t closure_depth) alloc_except;
 };
 
 }
