--- test/interpreter_scripts/in/namespace.txt
@@ -1,15 +1,0 @@
-namespace NAMESPACE {
-    test = "TEST"
-    //assert( NAMESPACE.test = "TEST" ) //DO THIS
-    
-    namespace inner {
-        test = "inner"
-    }
-}
-
-test = "OUTER"
-assert(test = "OUTER")
-assert( NAMESPACE.test = "TEST" )
-//assert( NAMESPACE.inner.test = "inner" ) //DO THIS
-
-print("namespace.txt PASS")
